Result Text = 'C' updates
A very small minority of extracted results are missing numerical results. Upon investigation in EMS a result of 'C' is allowed but this cannot be converted to numeric for EnMoDS so import fails. In cases where the EMS result is 'C' the exported result should be null and the detection condition should be set to 'NOT_SAMPLED' which is acceptable to EnMoDS. All other cases of missing result values must be filtered out from the final exported output.

 Example of the issue location ID E306061 date range 2019-11-4 to 2019-11-12 parameter code TEMF.